+++ shm_init_create.c
@@ -15,7 +15,7 @@ int shm_init_create(
 	key_t	shm_key,		// INPUT:   Keyword for Shared Memory
 	size_t	shm_size,		// INPUT:   Size of the SHM [bytes]
 	int		*shrd_id,		// OUTPUT:  Pointer to the Shared Memory ID 
-	int		*shm_ptr)		// OUTPUT:  Pointer to the Top of SHM
+	long	*shm_ptr)		// OUTPUT:  Pointer to the Top of SHM
 {
 /*
 -------------------------------------------------- OPEN SHARED MEMORY
@@ -27,7 +27,7 @@ int shm_init_create(
 		return(-1);
 	}
 
-	*shm_ptr = (int)shmat( *shrd_id, NULL, 0);
+	*shm_ptr = (long)shmat( *shrd_id, NULL, 0);
 	memset( (void *)*shm_ptr, 0, shm_size );
 
 	/*-------- ENDING --------*/
